Set charity default currency to THB
For now THB will be the only supported currency but this may change in the
future and it's why we don't set in as a default in the database. This
re-enforce the fact that all charities must be created using
app.create_charity(...) instead of the default rails way
using Charity.create(...).