Add the basic donation feature
In the view, interfaces are now provided to

  a) select a charity to which to donate to
  b) to enter an amount in thai baht
  c) to enter a credit card

The credit card is powered by Omise and will insert an omise_token into a hidden
field in the form. The amount must be greater than 20 THB. And a charity must be
selected for the action to work. If any of those information isn't present or
is not valid, the form will be re-rendered with a failure message. In case
everything went fine, one will be redirected to the same page with a success
message.